+++ SBBDesignSystemMobileSwiftUI/Views/SBBModalView/SBBModalView.swift
@@ -132,7 +132,7 @@ public struct SBBModalView<Content>: View where Content: View {
                     }
                     title
                         .padding(.top, 7)
-                        .padding(.leading, 16)
+                        .padding(.leading, 8)
                         .sbbFont(.large_light)
                         .accessibility(addTraits: .isHeader)
                         .accessibilitySortPriority(2)