+++ SBBDesignSystemMobileSwiftUI/Views/SBBTextField.swift
@@ -22,7 +22,8 @@ public struct SBBTextField: View {
     
     @Environment(\.isEnabled) private var isEnabled
     @Binding private var text: String
-    @State private var isEditing = false
+    @Binding var isEditing: Bool
+    @State private var isFocused = false
     let label: String?
     let error: String?
     let additionalAccessibilityText: String?
@@ -38,12 +39,14 @@ public struct SBBTextField: View {
         - text: Sets the user-modifiable text state.
         - label: An optional label displayed instead of the text (if the text is empty) or above the text (if not empty).
         - error: An optional label used to display errors with user input. Displayed below the text in red color and adding a red separator line below the entire SBBTextField.
+        - additionalAccessibilityText: An optional string appended at the end of the accessibility label.
         - icon: An optional Image to be shown on the leading edge of the SBBTextField.
         - showBottomLine: Shows or hides a separator line at the bottom of the View (typically only false for last elements in a List).
         - showClearButtonWhenEditing: Shows or hides the clear button when editing a text.
         - boxed: shows the Textfield inside a white box when enabled, and with a clear background when disabled (default). Clear background does not work inside SBBFormGroup.
+        - isEditing: a binding to get the state of the TextField (whether of not it is editing)
      */
-    public init(text: Binding<String>, label: String? = nil, error: String? = nil, additionalAccessibilityText: String? = nil, icon: Image? = nil, showBottomLine: Bool = true, showClearButtonWhenEditing: Bool = true, boxed: Bool = false) {
+    public init(text: Binding<String>, label: String? = nil, error: String? = nil, additionalAccessibilityText: String? = nil, icon: Image? = nil, showBottomLine: Bool = true, showClearButtonWhenEditing: Bool = true, boxed: Bool = false, isEditing: Binding<Bool>? = nil) {
         self._text = text
         if let label = label {
             self.label = NSLocalizedString(label, comment: "")
@@ -64,13 +67,14 @@ public struct SBBTextField: View {
         self.showBottomLine = showBottomLine
         self.showClearButtonWhenEditing = showClearButtonWhenEditing
         self.boxed = boxed
+        self._isEditing = isEditing != nil ? isEditing! : .constant(false)
     }
     
     private var bottomLineColor: Color {
         if error != nil {
             return .sbbColor(.red)
         }
-        switch (isEnabled, isEditing, showBottomLine) {
+        switch (isEnabled, isFocused, showBottomLine) {
         case (true, true, true):
             return .sbbColor(.textBlack)
         case (true, true, false):
@@ -100,7 +104,7 @@ public struct SBBTextField: View {
                     .accessibility(hidden: true)
             }
             VStack(alignment: .leading, spacing: 0) {
-                Group {
+                HStack {
                     if let label = label {
                         VStack(alignment: .leading, spacing: 4) {
                             if !text.isEmpty {
@@ -110,42 +114,40 @@ public struct SBBTextField: View {
                                     .opacity(text.isEmpty ? 0.0 : 1.0)
                                     .accessibility(hidden: true)
                             }
-                            TextField("", text: $text, onEditingChanged: { editing in
+                            TextField("\(label)", text: $text, onEditingChanged: { editing in
                                 DispatchQueue.main.async {
                                     withAnimation {
                                         self.isEditing = editing
+                                        self.isFocused = editing
                                     }
                                 }
                             })
                                 .modifier(TextFieldPlaceholderCustomColorStyle(showPlaceHolder: text.isEmpty, placeholder: label))
                                 .sbbFont(.medium_light)
                                 .accessibility(label: Text(accessibilityText))
-                                .accessibility(value: Text(""))
                         }
                     } else {
                         TextField("", text: $text, onEditingChanged: { editing in
                             DispatchQueue.main.async {
                                 withAnimation {
                                     self.isEditing = editing
+                                    self.isFocused = editing
                                 }
                             }
                         })
                             .sbbFont(.medium_light)
                     }
+                    
+                    if showClearButtonWhenEditing && isFocused && !text.isEmpty {
+                        Button(action: emptyText) {
+                            Image(sbbIcon: .cross_small)
+                                .accessibility(label: Text("Delete input".localized))
+                        }
+                            .buttonStyle(SBBIconButtonStyle(size: .small))
+                            .padding(.trailing, 16)
+                    }
                 }
                     .frame(minHeight: 48)
-                    .overlay(
-                        Group {
-                            if showClearButtonWhenEditing && isEditing && !text.isEmpty {
-                                Button(action: emptyText) {
-                                    Image(sbbIcon: .cross_small)
-                                        .accessibility(label: Text("Delete input".localized))
-                                }
-                                    .buttonStyle(SBBIconButtonStyle(size: .small))
-                                    .padding(.trailing, 16)
-                            }
-                        }
-                    , alignment: .trailing)
                 if let error = error {
                     Text(error)
                         .font(.sbbLight(size: 10))
@@ -153,7 +155,6 @@ public struct SBBTextField: View {
                         .padding(.bottom, 8)
                 }
             }
-            .animation(.linear)
             .background(bottomLineColor.frame(height: 1), alignment: .bottom)
         }
             .padding(.leading, 16)
@@ -168,6 +169,7 @@ public struct SBBTextField: View {
                         EmptyView()
                     }
                 } , alignment: .center)
+            .animation(.linear)
     }
     
     private func emptyText() {
@@ -182,7 +184,6 @@ public struct SBBTextField: View {
         if let error = error {
             text.append("\(error). ")
         }
-        text.append(self.text)
         
         if let additionalAccessibilityText = additionalAccessibilityText {
             text.append("\(additionalAccessibilityText). ")